+++ app/src/test/java/woowacourse/payments/ui/format/UniformlySeparatingVisualTransformationTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.formatter
+package woowacourse.payments.ui.format
 
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.input.TransformedText
